+++ crates/store/src/lib.rs
@@ -8,12 +8,7 @@ pub mod genesis;
 mod server;
 pub mod state;
 
-pub use accounts::{
-    AccountTreeStorage,
-    AccountTreeWithHistory,
-    HistoricalError,
-    InMemoryAccountTree,
-};
+pub use accounts::{AccountTreeWithHistory, HistoricalError, InMemoryAccountTree};
 pub use genesis::GenesisState;
 pub use server::{DataDirectory, Store};