+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.13.0 (TBD)
 
 - [BREAKING] Renamed `ProxyWorkerStatus::address` to `ProxyWorkerStatus::name` ([#1348](https://github.com/0xMiden/miden-node/pull/1348)).
+- Remove `trait AccountTreeStorage` ([#1352](https://github.com/0xMiden/miden-node/issues/1352)).
 
 ## v0.12.1 (2025-11-08)